Add the ability to specify the return file type
I saw this in another glob library and I thought it could be useful
here too. Basically you will now be able to specify which file types
you want to return from the glob. So if you only want to glob
directories or regular files that should work now. You can also
specify that you want everything as well.

I've decided to ignore symlinks entirely here. I'm not sure if that's
the right approach but it seems like the least surprising thing to do.
I guess we'll see if anyone has a strong opinion about that.

The method here is recursive but not tail-recursive. I don't think that
will be a problem though since the number of stackframes will in the
worst-case scenario equal the number of path components which shouldn't
end up being that large on any platform. It's also simpler to write
that sort of algorithm here.

I should also mention that the logic here is similar to that found in
the `simplifile.get_files` method found in simplifile v2.1.